moved testUtils to utils folder
